==> project-5/1_Data.ipynb <==
###############
### IMPORTS ###
###############

import pandas as pd
import numpy as np

The original dataset can be found at: https://www.kaggle.com/rounakbanik/the-movies-dataset

These files were too large to upload onto GitHub so they had to be deleted. When downloading the flatfiles, please place them in 'brian_nguyen/data/' path to read in the CSV.

### Read in CSV files as data frames

df_movies = pd.read_csv('data/movies_metadata.csv')
print('Length of df_movies: ', len(df_movies))
print('Columns: ', df_movies.columns)
df_movies.sample(5)

df_credits = pd.read_csv('data/credits.csv')
print('Length of df_credits: ', len(df_credits))
df_credits.sample(5)

df_keywords = pd.read_csv('data/keywords.csv')
print('Length of df_keywords: ', len(df_keywords))
df_keywords.sample(5)

df_links = pd.read_csv('data/links.csv')
print('Length of df_keywords: ', len(df_keywords))
df_links.sample(5)

# Rename 'movieId' to 'id' to match column names
df_links = df_links.rename(columns={'movieId': 'id'})
df_links.sample(5)

### Merge dataframe

print('Datatype of df_movies[id]', type(df_movies['id'].loc[1]))
print('Datatype of df_credits[id]', type(df_credits['id'].loc[1]))
print('Datatype of df_keywords[id]', type(df_keywords['id'].loc[1]))
print('Datatype of df_links[id]', type(df_links['id'].loc[1]))

# Change all ['id'] columns to str (from int64)
# This lets me keep the rows that cannot be converted to int64
df_credits['id'] = df_credits['id'].astype(str)
df_keywords['id'] = df_keywords['id'].astype(str)
df_links['id'] = df_links['id'].astype(str)

print('Datatype of df_movies[id]', type(df_movies['id'].loc[1]))
print('Datatype of df_credits[id]', type(df_credits['id'].loc[1]))
print('Datatype of df_keywords[id]', type(df_keywords['id'].loc[1]))
print('Datatype of dflinks[id]', type(df_links['id'].loc[1]))

df_temp1 = df_movies.merge(df_credits, how='inner', on='id')

df_temp1.head(3)

df = df_temp1.merge(df_keywords, how='inner', on='id')

df

### Functions to Obtain Clean Director, Cast, and Keywords

from ast import literal_eval

features = ['cast', 'crew', 'keywords', 'genres']
for feature in features:
    df[feature] = df[feature].apply(literal_eval)

# Get the director's name from the crew feature. If director is not listed, return NaN
def get_director(x):
    for i in x:
        if i['job'] == 'Director':
            return i['name']
    return np.nan

# Returns the list top 3 elements or entire list; whichever is more.
def get_list(x):
    if isinstance(x, list):
        names = [i['name'] for i in x]
        #Check if more than 3 elements exist. If yes, return only first three. If no, return entire list.
        if len(names) > 5:
            names = names[:5]
        return names

    #Return empty list in case of missing/malformed data
    return []

# Define new director, cast, genres and keywords features that are in a suitable form.
df['director'] = df['crew'].apply(get_director)

features = ['cast', 'keywords', 'genres']
for feature in features:
    df[feature] = df[feature].apply(get_list)

df[['title', 'cast', 'director', 'keywords', 'genres']].head(3)

df

# Export merged dataframe
df.to_csv('data/dataframe_merged.csv', index=False)
