==> project-5/3_Recommender_Collaborative-SmallNoDask.ipynb <==
## Recommender: Collaborative Filtering

This notebook shows the construction of a recommender system purely built from collaborative filtering.

###############
### IMPORTS ###
###############

# Calculating SVD matrix is too large so use Dask
import dask
import dask.dataframe as dd
import dask.array as da

import numpy as np
import pandas as pd

from scipy.linalg import svd
from sklearn.metrics.pairwise import cosine_similarity

df_all = dd.read_csv('data/dataframe_merged.csv')

print('Shape of dataframe: ', df_all.shape)
print('Columns of dataframe: ', df_all.columns)

I am interested in looking at the ratings dataset, where users and their ratings are mapped to movieIds. I will also load in the titles, so that I can refer to this to map the index to movie title.

df_titles = pd.read_csv('data/dataframe_merged.csv', usecols=['title', 'id'])

df_titles

df_ratings = pd.read_csv('data/ratings_small.csv')

df_ratings

df_movie_user = df_ratings.pivot_table(index='userId', columns='movieId', values='rating')
print('Shape of df_movie_user', df_movie_user.shape)

# Transpose matrix, fill in NaN with means, then transpose again
df_movie_user_imp = df_movie_user.T.fillna(df_movie_user.mean(axis=1)).T

df_movie_user_imp

U, sigma, Vt = svd(df_movie_user_imp)

pd.DataFrame(Vt)

def get_recommends(itemID, Vt, num_recom=2):
    recs = []
    for item in range(Vt.T.shape[0]):
        if item != itemID:
            recs.append([item,np.dot(Vt.T[itemID],Vt.T[item])])
    final_rec = [i[0] for i in sorted(recs,key=lambda x: x[1],reverse=True)]
    return final_rec[:num_recom]

#### get_recommends(2, Vt, num_recom=10)
