==> project-5/3_Recommender_Metadata-Jaccard.ipynb <==
###############
### IMPORTS ###
###############

import pandas as pd
import numpy as np

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics.pairwise import cosine_similarity, pairwise_distances
import nltk

df_all = pd.read_csv('data/dataframe_merged.csv')

print('Shape of dataframe: ', df_all.shape)
print('Columns of dataframe: ', df_all.columns)

df = pd.read_csv('data/dataframe_merged.csv', usecols=['id', 'title', 'genres', 'cast', 'director'])

print('Shape of dataframe: ', df.shape)
print('Columns of dataframe: ', df.columns)

df

# This will join first and last names to a single string (and lowercase) so that they do not
# become split during the vectorization process
def clean_data(x):
    if isinstance(x, list):
        return [str.lower(i.replace(" ", "")) for i in x]
    else:
        #Check if director exists. If not, return empty string
        if isinstance(x, str):
            return str.lower(x.replace(" ", ""))
        else:
            return ''

# Apply clean_data function to your features.
features = ['cast', 'director']

for feature in features:
    df[feature] = df[feature].apply(clean_data)

df

# Clean the text further by only keeping alphanumerics
import re

def create_metasoup(x):
    string = ''.join(x['cast']) + ' ' + x['director']
    return re.sub(r'\W+', ' ', string)
df['metasoup'] = df.apply(create_metasoup, axis=1)

df['metasoup']

df['metasoup'].loc[3]

# Using .split() will tokenize the row
df['metasoup'].loc[3].split()

df['metasoup']

# Tokenize each row in df['metasoup']
df['split_metasoup'] = df['metasoup'].apply(lambda x: x.split())

# Convert datatype to set
df['split_metasoup'] = df['metasoup'].apply(lambda x: set(x))

# tokenize dataset
df['split_metasoup'] = df['metasoup'].apply(lambda x: set(nltk.ngrams(nltk.word_tokenize(x), n=1)))

df['split_metasoup'].loc[0]

df['split_metasoup'].loc[3024]

df['metasoup'].loc[0]

# Calculate jaccard distance between Toy Story and Toy Story II
1 - nltk.jaccard_distance(df['split_metasoup'].loc[0], df['split_metasoup'].loc[3024])

# Create a shorter dataset to test creating df_jaccard below
df_short = df.head(100)
df_short

# Write a function to calculte jaccard distance
# Without the try-except loop, an error occurs: 'cannot divide by 0', when creating df_jaccard below
# https://python.gotrained.com/nltk-edit-distance-jaccard-distance/#Jaccard_Distance
def calculate_jaccard_dist(metasoup_A, metasoup_B):
    try:
        jaccard = 1 - nltk.jaccard_distance(metasoup_A, metasoup_B)
        return jaccard
    except:
        return 0

# Create df_jaccard (which will be our jaccard_similarity table)
df_jaccard = pd.DataFrame()

df_jaccard = df.apply(lambda y: 
    df.apply(lambda x: calculate_jaccard_dist(y['split_metasoup'], x['split_metasoup']), axis=1)
                         ,axis=1)

df_jaccard

jaccard_similarity = df_jaccard.to_numpy()

np.save('cosine_similarity/jaccard_metadata.npy', jaccard_similarity)

df_jaccard = pd.DataFrame(df_jaccard, index=df['split_metasoup'], columns=df['split_metasoup']

df_jaccard

vectorizer = CountVectorizer(stop_words='english')
doc_word = vectorizer.fit_transform(df['metasoup'])

pd.DataFrame(doc_word)

# Compute the cosine similarity matrix from doc_word
cosine_sim = cosine_similarity(doc_word, doc_word)

pd.DataFrame(cosine_sim)

pairwise = doc_word * doc_word.T

pd.DataFrame(pairwise.toarray())

pd.DataFrame(cosine_sim)

# # Save cosine_sim array to use in hybrid recommendation system
# np.save('cosine_similarity/cos_metadata.npy', cosine_sim)

# Reset index of our dataframe and construct reverse mapping as before
indices = pd.Series(df.index, index=df['title'])

# Function that takes in movie title as input and outputs most similar movies
def get_recommendations(title, cosine_sim=cosine_sim):
    # Get the index of the movie that matches the title
    idx = indices[title]

    # Get the pairwsie similarity scores of all movies with that movie
    sim_scores = list(enumerate(cosine_sim[idx]))

    # Sort the movies based on the similarity scores
    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)

    # Get the scores of the 10 most similar movies
    sim_scores = sim_scores[1:11]

    # Get the movie indices
    movie_indices = [i[0] for i in sim_scores]

    # Return the top 10 most similar movies
    return df['title'].iloc[movie_indices]

get_recommendations('The Dark Knight Rises')

get_recommendations('Toy Story')
