==> project-5/3_Recommender_OverviewTagline_Doc2Vec.ipynb <==
## Recommender: Overview of Movie Synopsis

This notebook uses the 'overview' column, which is concise description of the movie synopsis, to make a recommender. It uses a TfidfVectorizer to capture infrequently used words and weigh them more heavily (in contrast to CountVectorizer).

###############
### IMPORTS ###
###############

import pandas as pd
import numpy as np

import matplotlib as plt
import seaborn as sns


import re
import string


from gensim.models.doc2vec import Doc2Vec, TaggedDocument
from sklearn.model_selection import train_test_split


import nltk
from nltk import word_tokenize
import nltk
from nltk import word_tokenize
from sklearn import svm
from nltk.tag import StanfordNERTagger
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from nltk.tokenize.treebank import TreebankWordDetokenizer
from nltk.stem import PorterStemmer
from nltk.stem import LancasterStemmer
from nltk.stem.snowball import SnowballStemmer
from nltk.stem import WordNetLemmatizer
from nltk.corpus import words
from nltk.tag import pos_tag
from textblob import TextBlob
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

# Reference: http://zwmiller.com/projects/nlp_pipeline.html
# Reference: https://github.com/ZWMiller/nlp_pipe_manager/blob/master/nlp_pipeline_manager/nlp_preprocessor.py
# Reference: https://towardsdatascience.com/a-practitioners-guide-to-natural-language-processing-part-i-processing-understanding-text-9f4abfd13e72

class nlp_pipe:
    
    # Initialize the class
    def __init__(self, vectorizer, stemmer, lemmatizer, tokenizer, dataframe, column='Title'):
        self.vectorizer = vectorizer
        self.tokenizer = tokenizer
        self.lemmatizer = lemmatizer
        self.stemmer = stemmer
        self.dataframe = dataframe
        self.column = column
        self.dataframe[self.column] = self.dataframe[self.column].apply(str)
    
    ######################################################################
    
    # Create a cleaning method (aka fit) that will use several functions in order
    def cleaner(self):
        self.vader_sentiment()
        self.dataframe = self._remove_numbers(self.dataframe, self.column)
        self.dataframe = self._punctuation(self.dataframe, self.column)
        self.dataframe = self._dropduplicates(self.dataframe, self.column)
        self.real_words() # Check if it's a real word and then remove if not
        #self.remove_single_letter() # Remove single letter words
        #self.autocorrect() # Takes a very long time to run
        self.tokenize_words()
        self.remove_short_headlines() # Remove headline if only one word
        #self.lemmatize_words()
        #self.stem_words()
        #self.named_entities()
        self.dataframe = self._join_words(self.dataframe, self.column)
        #self.dataframe[self.column] = self.dataframe[self.column].replace('', np.nan,)
        #self.dataframe.dropna(subset=[self.column], inplace=True)
    
    ########## Functions that 'cleaner' will call ##########
    @staticmethod
    def _remove_numbers(dataframe, column):       
        # Removes all words containing numbers
        remove_numbers = lambda x: re.sub('\w*\d\w*', '', x)
        dataframe[column] = dataframe[column].map(remove_numbers)
        return dataframe
        
    @staticmethod
    def _punctuation(dataframe, column):
        # Removes punctuation marks
        punc_lower = lambda x: re.sub('[^A-Za-z0-9]+', ' ', x)
        dataframe[column] = dataframe[column].map(punc_lower)
        return dataframe
        
    @staticmethod
    def _dropduplicates(dataframe, column):
        # Drop rows that have duplicate 'Titles'
        dataframe.drop_duplicates(subset=column, keep='first', inplace=True)
        return dataframe
    
    @staticmethod
    def _join_words(dataframe, column):
        # Joins words together with space (' ')--used after tokenization
        join_words = lambda x: ' '.join(x)
        dataframe[column] = dataframe[column].map(join_words)
        return dataframe
    
    def vader_sentiment(self):
        analyzer = SentimentIntensityAnalyzer()
        self.dataframe['Positive_Sentiment'] = self.dataframe.apply(lambda x: analyzer.polarity_scores(x[self.column])['pos'], axis=1)
        self.dataframe['Negative_Sentiment'] = self.dataframe.apply(lambda x: analyzer.polarity_scores(x[self.column])['neg'], axis=1)
        self.dataframe['Neutral_Sentiment'] = self.dataframe.apply(lambda x: analyzer.polarity_scores(x[self.column])['neu'], axis=1)
        self.dataframe['Compound_Sentiment'] = self.dataframe.apply(lambda x: analyzer.polarity_scores(x[self.column])['compound'], axis=1)
        
    def tokenize_words(self):
        self.dataframe[self.column] = self.dataframe.apply(lambda x: self.tokenizer(x[self.column]), axis=1)
    
    def stem_words(self):
        self.dataframe[self.column] = self.dataframe.apply(lambda x: [self.stemmer.stem(word) for word in x[self.column]], axis=1)
                                                           
    def lemmatize_words(self):
        self.dataframe[self.column] = self.dataframe.apply(lambda x: [self.lemmatizer.lemmatize(word) for word in x[self.column]], axis=1)
        
    def named_entities(self):
        st = StanfordNERTagger('/usr/share/stanford-ner/classifiers/english.all.3class.distsim.crf.ser.gz',
                               '/usr/share/stanford-ner/stanford-ner.jar',
                               encoding='utf-8')
        self.dataframe[self.column] = self.dataframe.apply(lambda x: st.tag(x[self.column]), axis=1)
        
    def real_words(self):
        # Removes words that are not within the nltk.corpus library
        words = set(nltk.corpus.words.words())
        self.dataframe[self.column] = self.dataframe.apply(lambda x: \
        " ".join(w for w in nltk.wordpunct_tokenize(x[self.column]) if w.lower() in words or not w.isalpha()), axis=1)
        
    def remove_single_letter(self):
        # Removes words that are 1 letter
        self.dataframe[self.column] = self.dataframe.apply(lambda x: ' '.join([w for w in x[self.column].split() if len(w)>2]), axis=1)
        
    def remove_short_headlines(self, min_length=3):
        # Removes headlines that are less than 3 words
        self.dataframe['headline_length'] = self.dataframe.apply(lambda x: len(x[self.column]), axis=1)
        self.dataframe = self.dataframe[self.dataframe['headline_length'] > min_length]
        self.dataframe = self.dataframe.drop(columns='headline_length')
        self.dataframe.reset_index(drop=True)
        
    def remove_headlines_specific_words(self):
        self.dataframe = self.dataframe[~self.dataframe[self.column].str.contains('onion')]
        self.dataframe = self.dataframe[~self.dataframe[self.column].str.contains('Onion')]
        
    def autocorrect(self):
        # Autocorrects words based on Levenshtein distance (takes __ minutes to run)
        self.dataframe[self.column] = self.dataframe.apply(lambda x: ''.join(TextBlob(x[self.column]).correct()), axis=1)

df_all = pd.read_csv('data/dataframe_merged.csv')

print('Shape of dataframe: ', df_all.shape)
print('Columns of dataframe: ', df_all.columns)

df = pd.read_csv('data/dataframe_merged_small.csv', usecols=['id', 'title', 'overview', 'tagline'])

print('Shape of dataframe: ', df.shape)
print('Columns of dataframe: ', df.columns)

df

# Replace NaN with empty strings
df['overview'] = df['overview'].replace(np.nan, '', regex=True)
df['tagline'] = df['tagline'].replace(np.nan, '', regex=True)

# Join [overview] and [keywords] together
# These two columns are synopsis-associated and it's sensible to join them together
df['overview_and_tagline'] = df['overview'] + df['tagline']

df['overview_and_tagline'].loc[0]

### Clean the text using NLP class functions

# Clean the text using nlp_pipelines class
nlp = nlp_pipe(dataframe = df,
               column = 'overview_and_tagline',
               tokenizer = nltk.word_tokenize,
               vectorizer = TfidfVectorizer(stop_words='english'),
               stemmer = SnowballStemmer("english"),
               lemmatizer = WordNetLemmatizer())

# nlp.cleaner()

df

### Make the word2vec model and train on df['overview_and_tagline']

train, test = train_test_split(df, test_size=.20, random_state=42)

#Tag train set
tagged_tr = [TaggedDocument(words=word_tokenize(_d.lower()),
tags = [str(i)]) for i, _d in enumerate(train.overview_and_tagline)]

#Tag test set
tagged_test = [TaggedDocument(words=word_tokenize(_d.lower()),
tags=[str(i)]) for i, _d in enumerate(test.overview_and_tagline)]

model = Doc2Vec(vector_size=100,
                window=5, 
                alpha=.025, 
                min_alpha=0.00025, 
                min_count=2, 
                dm=1, 
                workers=8)

model.build_vocab(tagged_tr)

epochs = range(100)
for epoch in epochs:
    print(f'Epoch {epoch+1}')
    model.train(tagged_tr,
                total_examples=model.corpus_count,
                epochs=model.epochs)
    # decrease the learning rate
    model.alpha -= 0.00025
    # fix the learning rate, no decay
    model.min_alpha = model.alpha
     
model.save('data/doc2vec.model')

X_train = np.array([model.docvecs[str(i)] for i in range(len(tagged_tr))])

y_train = train['id']

y_train

print(model.most_similar('warrior'))

# Find movies similar to query word
vec = model['warrior']
print(model.docvecs.most_similar([vec]))

new_vector = model.infer_vector(tokens)
model.docvecs.most_similar([new_vector]) #gives you top 10 document tags and their cosine similarity

vectorizer = TfidfVectorizer(stop_words='english')
doc_word = vectorizer.fit_transform(df['overview_and_keywords'])

# Compute the cosine similarity matrix from doc_word
cosine_sim = cosine_similarity(doc_word, doc_word)

# Look at the cosine_sim matrix
pd.DataFrame(cosine_sim)

# # Save cosine_sim array to use in hybrid recommendation system
np.save('similarity_matrix/cos_overview_small.npy', cosine_sim)

# Reset index of our dataframe and construct reverse mapping as before
indices = pd.Series(df.index, index=df['title'])

# Function that takes in movie title as input and outputs most similar movies
def get_recommendations(title, cosine_sim=cosine_sim):
    # Get the index of the movie that matches the title
    idx = indices[title]

    # Get the pairwsie similarity scores of all movies with that movie
    sim_scores = list(enumerate(cosine_sim[idx]))

    # Sort the movies based on the similarity scores
    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)

    # Get the scores of the 10 most similar movies
    sim_scores = sim_scores[1:11]

    # Get the movie indices
    movie_indices = [i[0] for i in sim_scores]

    # Return the top 10 most similar movies
    return df['title'].iloc[movie_indices]

get_recommendations('The Dark Knight Rises')

get_recommendations('Toy Story')
