==> project-5/3_Recommender_Surprise.ipynb <==
import pandas as pd
import numpy as np

from surprise import accuracy
from surprise import SVD, NMF
from surprise import Dataset
from surprise import Reader
from surprise import NormalPredictor
from surprise.model_selection import cross_validate
from surprise.model_selection import KFold
from surprise.model_selection import GridSearchCV

# Read in the ratings_small (user-item) dataframe
df_ratings = pd.read_csv('data/ratings_10kusers.csv')

df_all = pd.read_csv('data/dataframe_merged.csv', usecols=['id', 'title'])

df_ratings = df_ratings[df_ratings['movieId'].isin(df_all['id'])]

df_ratings.tail(10)

# Set the reader to have a rating_scale from 1-5 (default)
reader = Reader(rating_scale=(1, 5))

# The data only consists of userId, movieId, and rating
data = Dataset.load_from_df(df_ratings[['userId', 'movieId', 'rating']], reader)

# Use surprise package for a train-test split of 80-20
# Note that the train-test split will split by general rows, not specific users
from surprise.model_selection import train_test_split
trainset, testset = train_test_split(data, test_size=0.20)

# trainsetfull = data.build_full_trainset()

print('Number of users: ', trainset.n_users)
print('Number of movies (items): ', trainset.n_items, '\n')

trainset_iids = list(trainset.all_items())
iid_converter = lambda x: trainset.to_raw_iid(x)
trainset_raw_iids = list(map(iid_converter, trainset_iids))

trainset_raw_iids

from surprise import KNNWithMeans
my_k = 15
my_min_k = 5
my_sim_option = {
    'name':'cosine', 'user_based':False, 'verbose': False
    }
algo = KNNWithMeans(sim_options = my_sim_option)
algo.fit(trainset)

# Dataframe of all the movies voted for
pd.DataFrame(algo.sim)

# Same dataframe as algo.sim but the indices/columns are now movieId
df_cos_surprise = pd.DataFrame(algo.sim, index=trainset_raw_iids, columns=trainset_raw_iids)

df_cos_surprise

df_all

df_all = df_all.reset_index()
df_all.index = df_all.id

df_all

movieIdtoindex = df_all['index'].to_dict()

df_cos_surprise = df_cos_surprise.rename(index=movieIdtoindex, columns=movieIdtoindex)

df_cos_surprise

# Make a pandas dataframe of movie x movie length from df_all
# Fill in the values from matrix 'algo.sim'
# Set the diagonal to "1"
df_blank = pd.DataFrame(np.nan, range(1,len(df_all)), range(1,len(df_all)))
df_blank

df_blank = df_cos_surprise.combine_first(df_blank)

df_blank

# # Assign the diagonals as 1
# df_blank.values[[np.arange(df_blank.shape[0])]*2] = 1

np.fill_diagonal(df_blank.values, 1)
df_blank

df_bank = df_blank.fillna(0)

df_blank

np.save('cosine_similarity/cos_ratings.npy', df_blank)
